requirements.txt
browserstack-local
nose
